feat(ch21): AlgebraicSpecial mode + Increment/Decrement unary ops
AlgebraicSpecial (--token-mode algebraic-special):
  Same as Algebraic but keeps {-1, 0, 1} as their real token IDs.
  These structural constants (loop terminators, boolean outputs,
  multiplicative/additive identities) remain visible to the transformer
  so it can learn patterns like `X0 - 1 == 0` (reach-zero check) and
  `X0 * 1 == X0` (identity) without those being collapsed into opaque
  algebraic slots. All other integers become X0, X1, …

Increment / Decrement (RouteType 8 & 9):
  Unary prefix ops (++x → x+1, --x → x-1). Only op1_pos (the single
  operand, most-recent numeric token) is required; op2_pos is ignored.
  apply_math_op uses b (the "right" operand slot) as the operand.
  MathLayer::forward updated: binary guard relaxed to allow op2_pos=-1
  for unary ops. Training data for ++/-- deferred until current
  comparison runs finish (adding them changes kNumRouteTypes=10 and
  thus router weight dims, invalidating in-progress checkpoints).

Also fixes pre-existing -Wsign-conversion in numeric_router.cpp route_hard.

All 415 tests pass; zero warnings.

+++ src/nn/numeric_router.cpp
@@ -75,7 +75,7 @@ std::vector<RouteType> NumericRouter::route_hard(const autograd::Variable& x) co
     for (std::size_t t = 0; t < T_sz; ++t) {
         int best = 0;
         for (std::size_t k = 1; k < K_sz; ++k) {
-            if (mask_sp[t * K_sz + k] > mask_sp[t * K_sz + best])
+            if (mask_sp[t * K_sz + k] > mask_sp[t * K_sz + static_cast<std::size_t>(best)])
                 best = static_cast<int>(k);
         }
         routes.push_back(static_cast<RouteType>(best));
